+++ package.json
@@ -8,7 +8,7 @@
     "lint": "eslint ./src ./webpack.config.js -f table || true",
     "server": "babel-node src/server/server.js",
     "build": "NODE_ENV=production && webpack -p",
-    "test": "echo 0"
+    "test": "jest"
   },
   "repository": {
     "type": "git",
@@ -42,6 +42,8 @@
     "eslint-plugin-jsx-a11y": "^4.0.0",
     "eslint-plugin-react": "^6.9.0",
     "file-loader": "^0.10.0",
+    "jest": "^18.1.0",
+    "jest-cli": "^18.1.0",
     "npm-run-all": "^4.0.1",
     "webpack-dashboard": "^0.3.0",
     "webpack-dev-middleware": "^1.10.0",
@@ -54,14 +56,14 @@
     "bluebird": "^3.4.7",
     "body-parser": "^1.16.0",
     "dotenv": "^4.0.0",
+    "emoji-mart": "^0.3.7",
     "express": "^4.14.1",
     "morgan": "^1.8.0",
     "mysql": "^2.13.0",
     "react": "^15.4.2",
     "react-dom": "^15.4.2",
     "react-redux": "^5.0.2",
-    "react-router": "4.0.0-alpha.5",
-    "react-router-dom": "^4.0.0-beta.5",
+    "react-router-dom": "next",
     "redux": "^3.6.0",
     "redux-promise": "^0.5.3",
     "socket.io": "^1.7.2",
