+++ src/index.js
@@ -1,8 +1,17 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
+import { BrowserRouter as Router, Route } from 'react-router-dom'
 
-import Routes from './client/Routes';
 import App from './client/App';
 import Landing from './client/Landing';
+import Room from './client/Room';
 
-ReactDOM.render(<Routes />, document.getElementById('container'))
+ReactDOM.render(
+  <Router>
+      <div>
+        <Route exact path="/" component={App} />
+        <Route exact path="/room" component={Room} />
+      </div>
+  </Router>,
+  document.getElementById('container')
+)
